Add a lights driver that outputs the sextet lights protocol over a serial connection, instead of to a pipe. Also refactor the sextet lights driver to use a common utility file that the serial driver takes advantage of as well.
